refactored App.js, moving message formatters into separate file
